+++ packages/schema-derive/src/cw_serde.rs
@@ -4,28 +4,30 @@ pub fn cw_serde_impl(input: DeriveInput) -> DeriveInput {
     match input.data {
         syn::Data::Struct(_) => parse_quote! {
             #[derive(
-                serde::Serialize,
-                serde::Deserialize,
-                Clone,
-                Debug,
-                PartialEq,
-                schemars::JsonSchema
+                ::cosmwasm_schema::serde::Serialize,
+                ::cosmwasm_schema::serde::Deserialize,
+                ::std::clone::Clone,
+                ::std::fmt::Debug,
+                ::std::cmp::PartialEq,
+                ::cosmwasm_schema::schemars::JsonSchema
             )]
             #[allow(clippy::derive_partial_eq_without_eq)] // Allow users of `#[cw_serde]` to not implement Eq without clippy complaining
-            #[serde(deny_unknown_fields)]
+            #[serde(deny_unknown_fields, crate = "::cosmwasm_schema::serde")]
+            #[schemars(crate = "::cosmwasm_schema::schemars")]
             #input
         },
         syn::Data::Enum(_) => parse_quote! {
             #[derive(
-                serde::Serialize,
-                serde::Deserialize,
-                Clone,
-                Debug,
-                PartialEq,
-                schemars::JsonSchema
+                ::cosmwasm_schema::serde::Serialize,
+                ::cosmwasm_schema::serde::Deserialize,
+                ::std::clone::Clone,
+                ::std::fmt::Debug,
+                ::std::cmp::PartialEq,
+                ::cosmwasm_schema::schemars::JsonSchema
             )]
             #[allow(clippy::derive_partial_eq_without_eq)] // Allow users of `#[cw_serde]` to not implement Eq without clippy complaining
-            #[serde(deny_unknown_fields, rename_all = "snake_case")]
+            #[serde(deny_unknown_fields, rename_all = "snake_case", crate = "::cosmwasm_schema::serde")]
+            #[schemars(crate = "::cosmwasm_schema::schemars")]
             #input
         },
         syn::Data::Union(_) => panic!("unions are not supported"),
@@ -47,15 +49,16 @@ mod tests {
 
         let expected = parse_quote! {
             #[derive(
-                serde::Serialize,
-                serde::Deserialize,
-                Clone,
-                Debug,
-                PartialEq,
-                schemars::JsonSchema
+                ::cosmwasm_schema::serde::Serialize,
+                ::cosmwasm_schema::serde::Deserialize,
+                ::std::clone::Clone,
+                ::std::fmt::Debug,
+                ::std::cmp::PartialEq,
+                ::cosmwasm_schema::schemars::JsonSchema
             )]
             #[allow(clippy::derive_partial_eq_without_eq)]
-            #[serde(deny_unknown_fields)]
+            #[serde(deny_unknown_fields, crate = "::cosmwasm_schema::serde")]
+            #[schemars(crate = "::cosmwasm_schema::schemars")]
             pub struct InstantiateMsg {
                 pub verifier: String,
                 pub beneficiary: String,
@@ -73,15 +76,16 @@ mod tests {
 
         let expected = parse_quote! {
             #[derive(
-                serde::Serialize,
-                serde::Deserialize,
-                Clone,
-                Debug,
-                PartialEq,
-                schemars::JsonSchema
+                ::cosmwasm_schema::serde::Serialize,
+                ::cosmwasm_schema::serde::Deserialize,
+                ::std::clone::Clone,
+                ::std::fmt::Debug,
+                ::std::cmp::PartialEq,
+                ::cosmwasm_schema::schemars::JsonSchema
             )]
             #[allow(clippy::derive_partial_eq_without_eq)]
-            #[serde(deny_unknown_fields)]
+            #[serde(deny_unknown_fields, crate = "::cosmwasm_schema::serde")]
+            #[schemars(crate = "::cosmwasm_schema::schemars")]
             pub struct InstantiateMsg {}
         };
 
@@ -101,15 +105,16 @@ mod tests {
 
         let expected = parse_quote! {
             #[derive(
-                serde::Serialize,
-                serde::Deserialize,
-                Clone,
-                Debug,
-                PartialEq,
-                schemars::JsonSchema
+                ::cosmwasm_schema::serde::Serialize,
+                ::cosmwasm_schema::serde::Deserialize,
+                ::std::clone::Clone,
+                ::std::fmt::Debug,
+                ::std::cmp::PartialEq,
+                ::cosmwasm_schema::schemars::JsonSchema
             )]
             #[allow(clippy::derive_partial_eq_without_eq)]
-            #[serde(deny_unknown_fields, rename_all = "snake_case")]
+            #[serde(deny_unknown_fields, rename_all = "snake_case", crate = "::cosmwasm_schema::serde")]
+            #[schemars(crate = "::cosmwasm_schema::schemars")]
             pub enum SudoMsg {
                 StealFunds {
                     recipient: String,